requirements.txt
backports.ssl-match-hostname==3.4.0.2
docker-py==1.2.3
gevent==1.0.2
greenlet==0.4.7
requests==2.7.0
six==1.9.0
websocket-client==0.32.0
pika
redis
